api/requirement.txt
asgiref
Django
django-cleanup
django-cors-headers
djangorestframework
Pillow
pytz
sqlparse
